Fix compiler warning on GNU/Linux
pcsc_scan.c: In function ‘displayChangedStatus’:
pcsc_scan.c:496:34: warning: format ‘%d’ expects argument of type ‘int’, but argument 4 has type ‘DWORD’ {aka ‘long unsigned int’} [-Wformat=]
  496 |                 printf("%d: %s, %d, 0x%04X -> 0x%04X",
      |                                 ~^
      |                                  |
      |                                  int
      |                                 %ld
......
  499 |                         r.dwEventState >> 16,
      |                         ~~~~~~~~~~~~~~~~~~~~
      |                                        |
      |                                        DWORD {aka long unsigned int}

+++ pcsc_scan.c
@@ -496,7 +496,7 @@ static void displayChangedStatus(SCARD_READERSTATE rgReaderStates[], int count)
 		printf("%d: %s, %d, 0x%04X -> 0x%04X",
 			i,
 			r.szReader,
-			r.dwEventState >> 16,
+			(int)(r.dwEventState >> 16),
 			(int)(r.dwCurrentState & 0xFFFF),
 			(int)(r.dwEventState & 0xFFFF));
 		for (int b=0; b<11; b++)
